requirements.txt
appdirs==1.4.4
attrs==20.1.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==19.9.0
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
pathspec==0.8.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
regex==2020.7.14
requests==2.24.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.1.0
